Rename otel.ts to rpc-tracing.ts
The module only holds the connect-rpc client tracing interceptor, not
general OTel setup (that's telemetry.ts) — name it for what it does.
