+++ packages/bsky/src/bsync.ts
@@ -9,8 +9,8 @@ import {
   type ConnectTransportOptions,
   createConnectTransport,
 } from '@connectrpc/connect-node'
-import { tracingInterceptor } from './otel.js'
 import { Service } from './proto/bsync_connect.js'
+import { tracingInterceptor } from './rpc-tracing.js'
 
 export type BsyncClient = PromiseClient<typeof Service>
